Extract chart tool helpers into shared module
